requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
scrapy==2.11.0
httpx==0.25.2
supabase==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.8
python-dotenv==1.0.0
sendgrid==6.11.0
email-validator==2.1.0
pandas==2.1.3
numpy==1.26.2
json5==0.9.14
APScheduler==3.10.4
schedule==1.2.0
click==8.1.7
tqdm==4.66.1
colorama==0.4.6
tenacity==8.2.3
pytz==2023.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
cryptography==41.0.7
python-json-logger==2.0.7
sentry-sdk==1.38.0
pytest==7.4.3
pytest-asyncio==0.21.1
redis==5.0.1
celery==5.3.4
Pillow==10.1.0
pytesseract==0.3.10
PyPDF2==4.0.1
pdfplumber==0.10.3
lxml==4.9.3
openpyxl==3.1.2
python-dateutil==2.8.2
python-multipart==0.0.6
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1